#57 added repositories folder (#59)
* #28-#29-#44-created-basic-lobby-functionality

* #57 added repositories folder
